[#21] Refactor: modify constants path to omit subpath
